requirement.txt
numpy
Pillow
mtcnn
keras_facenet
facenet
flask
split-folders